Rename coral-ui/FormField with TextField.
