+++ client/coral-admin/src/containers/Install/components/Steps/style.css
@@ -33,7 +33,7 @@
       margin: 30px 0;
     }
 
-    .formField {
+    .textField {
       text-align: left;
 
       label {
