+++ client/coral-ui/index.js
@@ -13,7 +13,7 @@ export {default as Icon} from './components/Icon';
 export {default as List} from './components/List';
 export {default as Item} from './components/Item';
 export {default as Card} from './components/Card';
-export {default as FormField} from './components/FormField';
+export {default as TextField} from './components/TextField';
 export {default as Success} from './components/Success';
 export {default as Pager} from './components/Pager';
 export {default as Wizard} from './components/Wizard';
